requirements.txt
attrs==23.2.0
certifi==2024.6.2
cfgv==3.4.0
distlib==0.3.8
exceptiongroup==1.2.1
filelock==3.15.4
h11==0.14.0
identify==2.5.36
idna==3.7
nodeenv==1.9.1
outcome==1.3.0.post0
platformdirs==4.2.2
pre-commit==3.7.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
wsproto==1.2.0
